refactored utils out of component code
