+++ src/app/layout.tsx
@@ -1,12 +1,12 @@
 import type { Metadata } from 'next';
 import localFont from 'next/font/local';
 import './globals.css';
-import { Inter } from '@next/font/google';
+// import { Inter } from '@next/font/google';
 
-const inter = Inter({
-    subsets: ['latin'],
-    variable: '--font-inter',
-});
+// const inter = Inter({
+//     subsets: ['latin'],
+//     variable: '--font-inter',
+// });
 
 const geistSans = localFont({
     src: './fonts/GeistVF.woff',
@@ -32,7 +32,7 @@ export default function RootLayout({
     return (
         <html lang="en">
             <body
-                className={`${inter.variable} ${geistSans.variable} ${geistMono.variable} antialiased`}
+                className={`${geistSans.variable} ${geistMono.variable} antialiased`}
             >
                 {children}
             </body>
