+++ src/components/Chart/utils/schema.ts
@@ -1,0 +1,4 @@
+export interface SignalData {
+    time: string;
+    signals: number[];
+}
